Clean up tests
This renames two of the test files to names that better reflect what they're
testing.

I also went through and adjusted the clean() tests to make it easier to read
them and verify the input and output. Granted, this is highly subjective and
specific to me, but I'm hoping the end result is that the tests are more uniform
in structure and easier to read for others, too.
